+++ backends/freecad/gui/freecad_bolts.py
@@ -20,26 +20,35 @@ bolts_path = dirname(__file__)
 try:
 	from PySide import QtCore, QtGui
 	from FreeCADGui import PySideUic as uic
-	Ui_BoltsWidget,QBoltsWidget = uic.loadUiType(join(bolts_path,'bolts_widget.ui'))
-	Ui_ValueWidget,QValueWidget = uic.loadUiType(join(bolts_path,'value_widget.ui'))
-	Ui_BoolWidget,QBoolWidget = uic.loadUiType(join(bolts_path,'bool_widget.ui'))
-	Ui_TableIndexWidget,QTableIndexWidget = uic.loadUiType(join(bolts_path,'tableindex_widget.ui'))
-	Ui_PropertyWidget,QPropertyWidget = uic.loadUiType(join(bolts_path,'property_widget.ui'))
+	Ui_BoltsWidget,_QBoltsWidget = uic.loadUiType(join(bolts_path,'bolts_widget.ui'))
+	Ui_ValueWidget,_QValueWidget = uic.loadUiType(join(bolts_path,'value_widget.ui'))
+	Ui_BoolWidget,_QBoolWidget = uic.loadUiType(join(bolts_path,'bool_widget.ui'))
+	Ui_TableIndexWidget,_QTableIndexWidget = uic.loadUiType(join(bolts_path,'tableindex_widget.ui'))
+	Ui_PropertyWidget,_QPropertyWidget = uic.loadUiType(join(bolts_path,'property_widget.ui'))
+
+	from PySide.QtGui import QDockWidget as QBoltsWidget
+	from PySide.QtGui import QWidget as QValueWidget
+	from PySide.QtGui import QWidget as QBoolWidget
+	from PySide.QtGui import QWidget as QTableIndexWidget
+	from PySide.QtGui import QWidget as QPropertyWidget
+
 	from PySide.QtCore import Slot
 	def unpack(x):
 		return x
 except ImportError:
 	from PyQt4 import QtGui, QtCore
 	from bolts_widget import Ui_BoltsWidget
-	from PyQt4.QtGui import QDockWidget as QBoltsWidget
 	from value_widget import Ui_ValueWidget
-	from PyQt4.QtGui import QWidget as QValueWidget
 	from bool_widget import Ui_BoolWidget
-	from PyQt4.QtGui import QWidget as QBoolWidget
 	from tableindex_widget import Ui_TableIndexWidget
-	from PyQt4.QtGui import QWidget as QTableIndexWidget
 	from property_widget import Ui_PropertyWidget
+
+	from PyQt4.QtGui import QDockWidget as QBoltsWidget
+	from PyQt4.QtGui import QWidget as QValueWidget
+	from PyQt4.QtGui import QWidget as QBoolWidget
+	from PyQt4.QtGui import QWidget as QTableIndexWidget
 	from PyQt4.QtGui import QWidget as QPropertyWidget
+
 	from PyQt4.QtCore import pyqtSlot as Slot
 	def unpack(x):
 		return x.toPyObject()
@@ -328,6 +337,9 @@ class BoltsWidget(QBoltsWidget):
 
 		for key,tp in data.parameters.types.iteritems():
 			if tp in lengths:
+				if params[key] is None:
+					#A undefined value is not necessarily fatal
+					continue
 				revision = int(FreeCAD.Version()[2].split()[0])
 				if revision >= 2836:
 					params[key] = FreeCAD.Units.parseQuantity("%g %s" %
@@ -337,9 +349,13 @@ class BoltsWidget(QBoltsWidget):
 						(params[key], lengths[tp]))
 
 		#add part
-		base = self.freecad.getbase[data.id]
-		add_part(base,params,FreeCAD.ActiveDocument)
-		FreeCADGui.SendMsgToActiveView("ViewFit")
+		try:
+			base = self.freecad.getbase[data.id]
+			add_part(base,params,FreeCAD.ActiveDocument)
+			FreeCADGui.SendMsgToActiveView("ViewFit")
+		except Exception as e:
+			FreeCAD.Console.PrintMessage(e)
+			QtGui.QErrorMessage(self).showMessage("An error occured when trying to add the part: %s\nParameter Values: %s" % (e,params))
 
 	@Slot()
 	def on_partsTree_itemSelectionChanged(self):