+++ bolttools/doc/changes.rst
@@ -17,6 +17,8 @@ Changes of the BOLTS specifications
 * added syntax to specify connectors for OpenSCAD
 * added description field to Parameter element
 * added table2 2 dimensional lookup tables
+* make types mandatory
+* distinguish between dimension and connector drawings